Lambda file
Lambda function to monitor S3 Uri. Whenever you uploads data into the S3 storage, the Lambda function should capture the details and send out an email to email-id displaying, S3 Uri, Object Name, Object Size, Object type and In addition to the above, the program should create a thumbnail and store it in a different uri in case the user uploads an image.